Merge room.js and rooms.js view models into main.js
Some modules, namely sync, player and playlist, must now be invoked as functions in order to instantiate them. This is because those modules have initialization behavior, e.g. the sync module opening a socket connection, and those behaviors should not execute just because the main module is loaded into the browser.